requirements.txt
robin-stocks
openai
requests
beautifulsoup4
python-dotenv
